fix(sequencing): store read depth as stated, not rounded to a whole number
REPORT02 finding 2. Three analyses held a depth of 0. The source agrees with
the zero in two of them, but the third states 0,38 — a sample that aligned and
then produced no usable coverage. Rounding on the way in made it
indistinguishable from a sample that produced nothing at all, which is exactly
the distinction someone judging a sample's usability needs.

MedianReadDepth becomes a double and is read with Number rather than Int32, so
524,81 arrives as 524.81. ObservedReadLength stays an integer: a read is not
sequenced two thirds of the way.

Worth noting for the uploader: FAIR Genomes types the catalogue's
"Median read depth" as an Integer, so whoever maps to it will have to round.
That belongs at the catalogue boundary rather than here — a source API should
keep what the source said and let the consumer lose precision deliberately.

(cherry picked from commit 3763bc53b65f1da60217d1c5a9604a5aca3f831c)

+++ src/SequencingApi/SequencingApi.Domain/QualityMetrics.cs
@@ -35,13 +35,22 @@ public sealed record QualityMetrics : ValueObject
     /// median — MMCI's pipeline states only a mean over the region of interest, and that is what
     /// lands here, exactly as the previous uploader did.
     /// </summary>
-    public int? MedianReadDepth { get; init; }
+    /// <remarks>
+    /// Fractional, though the catalogue's field is an integer. The sources state depths like
+    /// <c>524,81</c> and <c>0,38</c>, and rounding on the way in makes a sample that was sequenced
+    /// too shallowly to use indistinguishable from one that produced no coverage at all. A source
+    /// service should keep what the source said and leave rounding to whoever needs a whole number.
+    /// </remarks>
+    public double? MedianReadDepth { get; init; }
 
-    /// <summary>Read length the run actually achieved, as opposed to the length it was set up for.</summary>
+    /// <summary>
+    /// Read length the run actually achieved, as opposed to the length it was set up for. A whole
+    /// number of bases, unlike the depth above — a read is not sequenced two thirds of the way.
+    /// </summary>
     public int? ObservedReadLength { get; init; }
 
     public static ErrorOr<QualityMetrics> Create(
-        int? medianReadDepth = null,
+        double? medianReadDepth = null,
         int? observedReadLength = null)
     {
         if (medianReadDepth is < 0)
